+++ HousecatEngine/src/editor/Editor.h
@@ -80,7 +80,7 @@ public:
 	void Render();
 	void Destroy();
 
-	void AdjustZoom(int direction);
+	void AdjustZoom(double direction);
 
 	bool IsRunning() const { return isRunning; }
 	void SetRunning(bool running) { isRunning = running; }
